+++ src/java/arjdbc/jdbc/RubyJdbcConnection.java
@@ -2321,7 +2321,7 @@ public class RubyJdbcConnection extends RubyObject {
             }
             return array;
         }
-        finally { value.free(); }
+        finally { if ( value != null ) value.free(); }
     }
 
     protected IRubyObject xmlToRuby(final ThreadContext context,
@@ -2329,9 +2329,11 @@ public class RubyJdbcConnection extends RubyObject {
         throws SQLException {
         final SQLXML xml = resultSet.getSQLXML(column);
         try {
+            if ( xml == null || resultSet.wasNull() ) return runtime.getNil();
+
             return RubyString.newUnicodeString(runtime, xml.getString());
         }
-        finally { xml.free(); }
+        finally { if ( xml != null ) xml.free(); }
     }
 
     protected void setStatementParameters(final ThreadContext context,